Implement proper lexical scoping.
Discard Scope and ScopeChain, replacing them with LexicalEnvironment,
which in turn uses Declarative- and Object-EnvironmentRecords (like
described in ECMA spec).

Each JS::Function gets a pointer to an environment at its creation time,
and when the function is called, that environment (where the function
was defined) will be extended, not the environment where it was called
(like it was implemented before this commit).