Add 'this' parameter to all function calls.
Both JS::Function#call and JS::RubyFunction#call now expect first the
`this` parameter to be passed in and then the *arguments.

When called without context, the 'this' param will have to default to
the global object. Therefore added a Env::Lexical#global_object method
to easily access the global object.

The 'this' keyword now also works as it should.