Eliminate the use of VALUE[] wrapper.
Instead of wrapping each and every object inside VALUE[], from where
we in turn need to unwrap it each and every time, perform a little
bit of wrapping inside Runtime::Reference object in places where we
actually need this special handling that the wrapping provides,
notably when setting the values of variables and object properties.

The code looks now much simpler.