Let all functions reference the environment.
So, they can access the builtin Object.prototype from the Env::Lexical
object instead instantiating it by themself, and therefore they can all
point to the same shared Object.prototype.

For this to work, the properties of a global object are created
in a separate method, so we can first construct the global environment,
and then pass it to the #init_in_environment, which creates Function
objects.

As another major change, the environment is the first argument of
Function now, and also a required one.